websocket_django/requirements.txt
Django==5.0.1
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0
redis==5.0.1
openai==1.35.3
python-dotenv==1.0.0